+++ src/Raven.CodeAnalysis/Binder/BinderFactory.cs
@@ -35,6 +35,7 @@ class BinderFactory
             ElseExpressionClauseSyntax elseClause => new LocalScopeBinder(parentBinder!),
             WhileStatementSyntax stmt => new BlockBinder(parentBinder!.ContainingSymbol, parentBinder!),
             WhilePatternStatementSyntax stmt => new BlockBinder(parentBinder!.ContainingSymbol, parentBinder!),
+            LoopStatementSyntax stmt => new BlockBinder(parentBinder!.ContainingSymbol, parentBinder!),
             ForStatementSyntax stmt => new BlockBinder(parentBinder!.ContainingSymbol, parentBinder!),
             FunctionStatementSyntax localFunc => new FunctionBinder(parentBinder!, localFunc),
             // Type declaration binders are created and cached by SemanticModel