+++ src/Raven.CodeAnalysis/BoundTree/BoundLoopStatement.cs
@@ -1,0 +1,11 @@
+namespace Raven.CodeAnalysis;
+
+internal partial class BoundLoopStatement : BoundStatement
+{
+    public BoundStatement Body { get; }
+
+    public BoundLoopStatement(BoundStatement body)
+    {
+        Body = body;
+    }
+}